调整系统目录和文件存储结构
