+++ demo03/src/main/webapp/WEB-INF/views/dengmi/queryMicai.jsp
@@ -1,0 +1,16 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: YowYouth
+  Date: 2019/3/29
+  Time: 15:06
+  To change this template use File | Settings | File Templates.
+--%>
+<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<html>
+<head>
+    <title>Title</title>
+</head>
+<body>
+查询谜材
+</body>
+</html>
