+++ demo03/src/main/webapp/WEB-INF/views/dengmi/sayDengmi.jsp
@@ -1,0 +1,16 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: YowYouth
+  Date: 2019/3/29
+  Time: 15:08
+  To change this template use File | Settings | File Templates.
+--%>
+<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<html>
+<head>
+    <title>Title</title>
+</head>
+<body>
+赏谜论谜
+</body>
+</html>
